feat: simplify implementation
